+++ CHANGELOG.md
@@ -13,8 +13,16 @@ for further information about branching and tagging conventions.
 
 - Rename PaymentV2 builder to Payment builder
 
+
+## [1.0.0-x.14] - 2019-02-21
+### Added
+- KeyServerCaller
+
+### Fixed
+- Wallet.fromEncrypted in cases when you provide seed only
+
 ## [1.0.0-x.13] - 2019-02-21
-### Adde
+### Added
 - Limits & bind external system ID unit tests
 
 ### Fixed